+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/devices/huami/amazfitbip/AmazfitBipCoordinator.java
@@ -68,4 +68,9 @@ public class AmazfitBipCoordinator extends HuamiCoordinator {
     public boolean supportsActivityTracks() {
         return true;
     }
+
+    @Override
+    public boolean supportsWeather() {
+        return true;
+    }
 }