+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/devices/huami/amazfitcor/AmazfitCorCoordinator.java
@@ -63,4 +63,9 @@ public class AmazfitCorCoordinator extends HuamiCoordinator {
     public boolean supportsHeartRateMeasurement(GBDevice device) {
         return true;
     }
+
+    @Override
+    public boolean supportsWeather() {
+        return true;
+    }
 }