+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/devices/jyou/TeclastH30Coordinator.java
@@ -63,7 +63,7 @@ public class TeclastH30Coordinator extends AbstractDeviceCoordinator {
     @Override
     public DeviceType getSupportedType(GBDeviceCandidate candidate) {
         String name = candidate.getDevice().getName();
-        if (name != null && name.startsWith("TECLAST_H30")) {
+        if (name != null && (name.startsWith("TECLAST_H30") || name.startsWith("TECLAST_H10"))) {
             return DeviceType.TECLASTH30;
         }
         return DeviceType.UNKNOWN;
@@ -82,6 +82,11 @@ public class TeclastH30Coordinator extends AbstractDeviceCoordinator {
     @Override
     public boolean supportsRealtimeData() {
         return true;
+    }
+
+    @Override
+    public boolean supportsWeather() {
+        return false;
     }
 
     @Override