+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/model/WeatherSpec.java
@@ -16,16 +16,106 @@
     along with this program.  If not, see <http://www.gnu.org/licenses/>. */
 package nodomain.freeyourgadget.gadgetbridge.model;
 
+import android.os.Parcel;
+import android.os.Parcelable;
+
+import java.util.ArrayList;
+
 // FIXME: document me and my fields, including units
-public class WeatherSpec {
+public class WeatherSpec implements Parcelable {
+    public static final Creator<WeatherSpec> CREATOR = new Creator<WeatherSpec>() {
+        @Override
+        public WeatherSpec createFromParcel(Parcel in) {
+            return new WeatherSpec(in);
+        }
+
+        @Override
+        public WeatherSpec[] newArray(int size) {
+            return new WeatherSpec[size];
+        }
+    };
     public int timestamp;
     public String location;
     public int currentTemp;
-    public int currentConditionCode;
+    public int currentConditionCode = 3200;
     public String currentCondition;
     public int todayMaxTemp;
     public int todayMinTemp;
-    public int tomorrowMaxTemp;
-    public int tomorrowMinTemp;
-    public int tomorrowConditionCode;
+    public ArrayList<Forecast> forecasts = new ArrayList<>();
+
+    public WeatherSpec() {
+
+    }
+
+    protected WeatherSpec(Parcel in) {
+        timestamp = in.readInt();
+        location = in.readString();
+        currentTemp = in.readInt();
+        currentConditionCode = in.readInt();
+        currentCondition = in.readString();
+        todayMaxTemp = in.readInt();
+        todayMinTemp = in.readInt();
+        in.readList(forecasts, Forecast.class.getClassLoader());
+    }
+
+    @Override
+    public int describeContents() {
+        return 0;
+    }
+
+    @Override
+    public void writeToParcel(Parcel dest, int flags) {
+        dest.writeInt(timestamp);
+        dest.writeString(location);
+        dest.writeInt(currentTemp);
+        dest.writeInt(currentConditionCode);
+        dest.writeString(currentCondition);
+        dest.writeInt(todayMaxTemp);
+        dest.writeInt(todayMinTemp);
+        dest.writeList(forecasts);
+    }
+
+    public static class Forecast implements Parcelable {
+        public static final Creator<Forecast> CREATOR = new Creator<Forecast>() {
+            @Override
+            public Forecast createFromParcel(Parcel in) {
+                return new Forecast(in);
+            }
+
+            @Override
+            public Forecast[] newArray(int size) {
+                return new Forecast[size];
+            }
+        };
+        public int minTemp;
+        public int maxTemp;
+        public int conditionCode;
+
+        public Forecast() {
+        }
+
+        public Forecast(int minTemp, int maxTemp, int conditionCode) {
+            this.minTemp = minTemp;
+            this.maxTemp = maxTemp;
+            this.conditionCode = conditionCode;
+        }
+
+        Forecast(Parcel in) {
+            minTemp = in.readInt();
+            maxTemp = in.readInt();
+            conditionCode = in.readInt();
+        }
+
+        @Override
+        public int describeContents() {
+            return 0;
+        }
+
+        @Override
+        public void writeToParcel(Parcel dest, int flags) {
+            dest.writeInt(minTemp);
+            dest.writeInt(maxTemp);
+            dest.writeInt(conditionCode);
+        }
+    }
 }