+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/service/devices/huami/HuamiDeviceEvent.java
@@ -24,8 +24,10 @@ public class HuamiDeviceEvent {
     public static final byte BUTTON_PRESSED = 0x04;
     public static final byte START_NONWEAR = 0x06;
     public static final byte CALL_REJECT = 0x07;
-    public static final byte CALL_ACCEPT = 0x09;
+    public static final byte FIND_PHONE_START = 0x08;
+    public static final byte CALL_IGNORE = 0x09;
     public static final byte ALARM_TOGGLED = 0x0a;
     public static final byte BUTTON_PRESSED_LONG = 0x0b;
     public static final byte TICK_30MIN = 0x0e; // unsure
+    public static final byte FIND_PHONE_STOP = 0x0f;
 }